+++ index.html
@@ -4,6 +4,7 @@
     <meta charset="UTF-8" />
     <link rel="icon" type="image/svg+xml" href="/vite.svg" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <script src="https://kit.fontawesome.com/bb8d2ebf9e.js" crossorigin="anonymous"></script>
     <title>Lyric Jam</title>
   </head>
   <body>
