+++ src/App.jsx
@@ -1,13 +1,14 @@
 import React from 'react';
 import { useEffect, useState } from 'react'
 import { GetSearchResult } from './API/APICalls.js';
-import YouTubeEmbed from './Components/YouTubeEmbed';
+import YouTubeEmbed from './components/YouTubeEmbed';
 import './App.css';
-import Header from './Header';
-import Hero from './Hero';
-import About from './About';
-import SearchResults from './SearchResults';
-import Facts from './Facts';
+import Header from './components/Header.jsx';
+import Hero from './components/Hero.jsx';
+import About from './components/About.jsx';
+import SearchResults from './components/SearchResults.jsx';
+import Facts from './components/Facts.jsx';
+import Footer from './components/Footer.jsx';
 
 function App() {
   const[results, setResults] = useState([]);
@@ -29,6 +30,7 @@ function App() {
       <SearchResults results={results}/>
       <Facts />
       <About />
+      <Footer />
     </React.Fragment>
   );
 }
