+++ src/Components/Footer.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+const Footer = () => {
+    return (
+        <footer className="bg-custom-red text-custom-gray text-center p-4 fixed bottom-0 min-w-full z-50">
+            <i class="fa-brands fa-square-facebook px-2 text-3xl"></i>
+            <i class="fa-brands fa-square-instagram px-2 text-3xl"></i>
+            <i class="fa-brands fa-square-snapchat px-2 text-3xl"></i>
+        </footer>
+    );
+};
+
+export default Footer;
